+++ application/controllers/admin/Profil.php
@@ -61,7 +61,9 @@ class Profil extends CI_Controller
                 $foto = $this->db->select('usr_foto')->where('usr_id', $_POST['profil_usr_id'])->get('system_users')->row()->usr_foto;
                 if (!empty($data_foto['file_name'])) {
                     if (!empty($foto)) {
-                        unlink($config['upload_path'] . $foto);
+                        if (file_exists($config['upload_path'] . $foto)) {
+                            unlink($config['upload_path'] . $foto);
+                        }
                     }
                 }
                 if (!empty($_POST['profil_usr_password'])) {
